# Improve performance of Order{By}{Descending}(...).First/Last