When there's no ThenBy, we can take a more optimized path that uses the TKey's comparer directly.  We already have a fast path for this case that converts the O(n log n) operation into O(n), but it employs a comparer that's much more complicated, and as that comparer is used in the inner loop, it makes a meaningful difference.

```C#
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;
using System.Runtime.InteropServices;

BenchmarkSwitcher.FromAssembly(typeof(Tests).Assembly).Run(args);

[MemoryDiagnoser(false)]
public class Tests
{
    [Params(8, 8000)]
    public int Count { get; set; }

    private List<double> _doubles;
    private List<int> _ints;
    private string[] _strings;

    [GlobalSetup]
    public void Setup()
    {
        _doubles = new(Enumerable.Range(-Count, Count * 2).Select(x => (double)x));
        new Random(42).Shuffle(CollectionsMarshal.AsSpan(_doubles));

        _ints = new(Enumerable.Range(-Count, Count * 2));
        new Random(42).Shuffle(CollectionsMarshal.AsSpan(_doubles));

        _strings = Enumerable.Range(-Count, Count * 2).Select(x => x.ToString()).ToArray();
        new Random(42).Shuffle(_strings);
    }

    [Benchmark] public double OrderByFirst_Double() => _doubles.OrderBy(x => x).First();
    [Benchmark] public double OrderLast_Double() => _doubles.Order().Last();

    [Benchmark] public int OrderByFirst_Int32() => _ints.OrderBy(x => x).First();
    [Benchmark] public int OrderLast_Int32() => _ints.Order().Last();

    [Benchmark] public string OrderByFirst_String() => _strings.OrderBy(x => x).First();
    [Benchmark] public string OrderLast_String() => _strings.Order().Last();
}
```

| Method              | Toolchain         | Count | Mean          | Ratio |
|-------------------- |------------------ |------ |--------------:|------:|
| OrderByFirst_Double | \main\corerun.exe | 8     |     113.16 ns |  1.00 |
| OrderByFirst_Double | \pr\corerun.exe   | 8     |      63.40 ns |  0.56 |
|                     |                   |       |               |       |
| OrderLast_Double    | \main\corerun.exe | 8     |     120.98 ns |  1.00 |
| OrderLast_Double    | \pr\corerun.exe   | 8     |      58.75 ns |  0.49 |
|                     |                   |       |               |       |
| OrderByFirst_Int32  | \main\corerun.exe | 8     |     112.58 ns |  1.00 |
| OrderByFirst_Int32  | \pr\corerun.exe   | 8     |      57.93 ns |  0.51 |
|                     |                   |       |               |       |
| OrderLast_Int32     | \main\corerun.exe | 8     |     106.44 ns |  1.00 |
| OrderLast_Int32     | \pr\corerun.exe   | 8     |      21.46 ns |  0.20 |
|                     |                   |       |               |       |
| OrderByFirst_String | \main\corerun.exe | 8     |     768.86 ns |  1.00 |
| OrderByFirst_String | \pr\corerun.exe   | 8     |     710.81 ns |  0.93 |
|                     |                   |       |               |       |
| OrderLast_String    | \main\corerun.exe | 8     |     684.53 ns |  1.00 |
| OrderLast_String    | \pr\corerun.exe   | 8     |     682.01 ns |  1.01 |
|                     |                   |       |               |       |
| OrderByFirst_Double | \main\corerun.exe | 8000  |  77,141.19 ns |  1.00 |
| OrderByFirst_Double | \pr\corerun.exe   | 8000  |  31,144.22 ns |  0.40 |
|                     |                   |       |               |       |
| OrderLast_Double    | \main\corerun.exe | 8000  |  71,845.74 ns |  1.00 |
| OrderLast_Double    | \pr\corerun.exe   | 8000  |  31,260.30 ns |  0.44 |
|                     |                   |       |               |       |
| OrderByFirst_Int32  | \main\corerun.exe | 8000  |  66,197.03 ns |  1.00 |
| OrderByFirst_Int32  | \pr\corerun.exe   | 8000  |  34,464.89 ns |  0.52 |
|                     |                   |       |               |       |
| OrderLast_Int32     | \main\corerun.exe | 8000  |  55,260.46 ns |  1.00 |
| OrderLast_Int32     | \pr\corerun.exe   | 8000  |     598.10 ns |  0.01 |
|                     |                   |       |               |       |
| OrderByFirst_String | \main\corerun.exe | 8000  | 835,626.65 ns |  1.00 |
| OrderByFirst_String | \pr\corerun.exe   | 8000  | 824,369.56 ns |  0.99 |
|                     |                   |       |               |       |
| OrderLast_String    | \main\corerun.exe | 8000  | 714,980.82 ns |  1.00 |
| OrderLast_String    | \pr\corerun.exe   | 8000  | 676,364.24 ns |  0.95 |

Fixes https://github.com/dotnet/runtime/issues/87921