@stephentoub 2024-01-26
> How do the benchmarks compare to using `MinBy` or `MaxBy`?

```C#
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;
using System.Runtime.InteropServices;

BenchmarkSwitcher.FromAssembly(typeof(Tests).Assembly).Run(args);

[MemoryDiagnoser(false)]
public class Tests
{
    [Params(1000)]
    public int Count { get; set; }

    private List<double> _doubles;
    private List<int> _ints;
    private string[] _strings;

    [GlobalSetup]
    public void Setup()
    {
        _doubles = new(Enumerable.Range(-Count, Count * 2).Select(x => (double)x));
        new Random(42).Shuffle(CollectionsMarshal.AsSpan(_doubles));

        _ints = new(Enumerable.Range(-Count, Count * 2));
        new Random(42).Shuffle(CollectionsMarshal.AsSpan(_doubles));

        _strings = Enumerable.Range(-Count, Count * 2).Select(x => x.ToString()).ToArray();
        new Random(42).Shuffle(_strings);
    }

    [Benchmark] public double OrderByLast_Double() => _doubles.OrderBy(x => x).Last();
    [Benchmark] public double OrderLast_Double() => _doubles.Order().Last();

    [Benchmark] public int OrderByLast_Int32() => _ints.OrderBy(x => x).Last();
    [Benchmark] public int OrderLast_Int32() => _ints.Order().Last();

    [Benchmark] public string OrderByLast_String() => _strings.OrderBy(x => x).Last();
    [Benchmark] public string OrderLast_String() => _strings.Order().Last();

    [Benchmark] public double MaxBy_Double() => _doubles.MaxBy(x => x);
    [Benchmark] public string MaxBy_String() => _strings.MaxBy(x => x);
    [Benchmark] public int MaxBy_Int32() => _ints.MaxBy(x => x);
}
```

| Method             | Count | Mean         | Error        | StdDev       | Allocated |
|------------------- |------ |-------------:|-------------:|-------------:|----------:|
| OrderByLast_Double | 1000  |  3,908.00 ns |    43.877 ns |    36.639 ns |      96 B |
| OrderLast_Double   | 1000  |  3,873.96 ns |    33.805 ns |    29.968 ns |      96 B |
| OrderByLast_Int32  | 1000  |  3,115.45 ns |     7.384 ns |     6.907 ns |      96 B |
| OrderLast_Int32    | 1000  |     88.72 ns |     0.331 ns |     0.309 ns |      32 B |
| OrderByLast_String | 1000  | 81,043.47 ns |   430.201 ns |   402.410 ns |      88 B |
| OrderLast_String   | 1000  | 81,615.21 ns |   804.978 ns |   752.977 ns |      88 B |
| MaxBy_Double       | 1000  |  3,187.21 ns |     3.413 ns |     3.025 ns |      40 B |
| MaxBy_String       | 1000  | 85,013.36 ns | 1,665.653 ns | 1,558.053 ns |      32 B |
| MaxBy_Int32        | 1000  |  3,805.06 ns |    13.573 ns |    12.032 ns |      40 B |